// MCP_Frontend-main/src/app/shared/models/customer-details.ts
export class CustomerDetail {
    userId: number;
    firstName: string;
    surName: string;
    dateOfBirth: Date;
    town: string;
    idType: string;
    idNumber: string;
    titleId: string;
    mobilePhone: string;
    idExpiryDate: Date;
    idIssueDt: Date;
    gender: string;
    outletNo: string;
    nokName: string;
    nokPhone: string;
    createDate: Date;
    photoBase64String: string;
    signatureBase64String: string;
}